fix(Postgres Node): Allow passing in arrays to JSON columns for insert
An array is technically JSON. Postgres allows for this and we should
enable this for the insert node.

We will need to get the `column` node parameters first and then call
values on it since the validator in core will not accept arrays to be
passed in for columns with JSON. As a result, we will use `pgp.as.json`
to handle validation. This also gives us the added benefit of converting
the values which are meant to be JSON into a format Postgres can
understand. We also need to enable raw mode since we will use
`pgp.helpers.concat` for the queries in `configureQueryRunners`.